requirements.txt
python-telegram-bot==13.1
requests==2.22.0